requirements.txt
Flask
Flask-WTF
openai
requests
python-dotenv
selenium
email-validator